+++ src/config/google.config.ts
