+++ src/index.ts
@@ -4,9 +4,9 @@ import express from "express";
 import createLogger from "./lib/wintson.logger";
 import jwtRoutes from "./routes/jwt.routes";
 import sessionRoutes from "./routes/session.routes";
-import setupGoogleRoutes from "./routes/setup-google-oath.routes";
-import setupSession from "./config/Session.config";
-import setupGoogleOath from "./config/GoogleOath.config";
+import setupGoogleRoutes from "./routes/google.routes";
+import setupSession from "./config/session.config";
+import setupGoogleOath from "./config/google.config";
 import jwtMiddleware from "./middlewares/jwt.middleware";
 import sessionMiddleware from "./middlewares/session.middleware";
 import googleAuthMiddleware from "./middlewares/googleAuth.middleware";
