+++ tests/index.test.ts
@@ -5,9 +5,9 @@ import { Config } from "../src/interfaces/config.interface";
 import createLogger from "../src/lib/wintson.logger";
 import jwtRoutes from "../src/routes/jwt.routes";
 import sessionRoutes from "../src/routes/session.routes";
-import setupGoogleRoutes from "../src/routes/setup-google-oath.routes";
-import setupSession from "../src/config/Session.config";
-import setupGoogleOath from "../src/config/GoogleOath.config";
+import setupGoogleRoutes from "../src/routes/google.routes";
+import setupSession from "../src/config/session.config";
+import setupGoogleOath from "../src/config/google.config";
 import jwtMiddleware from "../src/middlewares/jwt.middleware";
 import sessionMiddleware from "../src/middlewares/session.middleware";
 import googleAuthMiddleware from "../src/middlewares/googleAuth.middleware";
@@ -20,9 +20,9 @@ jest.mock("../src/lib/wintson.logger", () => () => ({
 }));
 jest.mock("../src/routes/jwt.routes");
 jest.mock("../src/routes/session.routes");
-jest.mock("../src/routes/setup-google-oath.routes");
-jest.mock("../src/config/Session.config");
-jest.mock("../src/config/GoogleOath.config");
+jest.mock("../src/routes/google.routes");
+jest.mock("../src/config/session.config");
+jest.mock("../src/config/google.config");
 jest.mock("../src/middlewares/jwt.middleware", () => jest.fn(() => (req: any, res: any, next: any) => next()));
 jest.mock("../src/middlewares/session.middleware", () => jest.fn(() => (req: any, res: any, next: any) => next()));
 jest.mock("../src/middlewares/googleAuth.middleware", () => jest.fn(() => (req: any, res: any, next: any) => next()));
